Crate penumbra_sdk_asset

Re-exports

pub use balance::Balance;

Modules

asset

Asset types and identifiers.

balance

Structs

EquivalentValue

An equivalent value in terms of a different numeraire.

EstimatedPrice

The estimated price of one asset in terms of another.

Value

ValueVar

Enums

ValueView

Represents a value of a known or unknown denomination.

Statics

STAKING_TOKEN_ASSET_ID

STAKING_TOKEN_DENOM