penumbra_sdk_asset::balance
Struct Balance

pub struct Balance {
    pub negated: bool,
    pub balance: BTreeMap<Id, Imbalance<NonZeroU128>>,
}

A Balance is a “vector of Values”, where some values may be required, while others may be provided. For a transaction to be valid, its balance must be zero.

Fields
negated: boolbalance: BTreeMap<Id, Imbalance<NonZeroU128>>
Implementations

impl Balance

pub fn zero() -> Self

Make a new, zero balance.

pub fn is_zero(&self) -> bool

Check if this balance is zero.

pub fn dimension(&self) -> usize

Find out how many distinct assets are represented in this balance.

pub fn required( &self, ) -> impl Iterator<Item = Value> + DoubleEndedIterator + FusedIterator + '_

Iterate over all the requirements of the balance, as Values.

pub fn provided( &self, ) -> impl Iterator<Item = Value> + DoubleEndedIterator + FusedIterator + '_

pub fn commit(&self, blinding_factor: Fr) -> Commitment

Commit to a Balance using a provided blinding factor.

This is like a vectorized Value::commit.